refactor(ctw.studio): remove client runtimes
