+++ ctw.studio/astro.config.mjs
@@ -1,5 +1,4 @@
 import { defineConfig } from 'astro/config';
-import svelte from '@astrojs/svelte';
 import tailwindcss from '@tailwindcss/vite';
 
 export default defineConfig({
@@ -10,7 +9,6 @@ export default defineConfig({
   build: {
     format: 'directory'
   },
-  integrations: [svelte()],
   vite: {
     plugins: [tailwindcss()]
   }
