+++ ctw.studio/signals/tests/demography.test.mjs
@@ -147,8 +147,8 @@ test('page carries substantive no-JS evidence and six accessible component table
   assert.match(html, /UN DESA Population Division · WPP 2024 · Medium variant/);
   assert.match(html, /No demographic burden score is created/i);
   assert.match(html, /Gross immigration and gross emigration remain visible/i);
-  assert.match(html, /href="\/signals\/" data-astro-reload(?:="true")?>Signals \//);
-  assert.match(html, /href="\/signals\/demography\/" aria-current="location" data-astro-reload/);
+  assert.match(html, /href="\/signals\/">Signals \//);
+  assert.match(html, /href="\/signals\/demography\/" aria-current="location"/);
 });
 
 test('renderer is dependency-free, uses baked JSON, handles missing data, and updater fails closed', () => {
