+++ ctw.studio/signals/tests/education.test.mjs
@@ -229,8 +229,8 @@ test('page has substantive no-JS content, semantic tables, components, and navig
   ]) assert.match(html, new RegExp(phrase, 'i'));
   assert.match(html, /<noscript>/);
   assert.match(html, /No capability composite is reported/i);
-  assert.match(html, /href="\/signals\/" data-astro-reload(?:="true")?>Signals \//);
-  assert.match(html, /href="\/signals\/ai-work\/" aria-current="location" data-astro-reload/);
+  assert.match(html, /href="\/signals\/">Signals \//);
+  assert.match(html, /href="\/signals\/ai-work\/" aria-current="location"/);
   assert.match(css, /@media \(max-width: 600px\)/);
 });
