+++ ctw.studio/signals/tests/science.test.mjs
@@ -70,8 +70,8 @@ test('science page contains substantive no-JS evidence, tables, sources, and com
   assert.match(html, /id="science-data-table"/);
   assert.match(html, /<noscript>/);
   assert.match(html, /Publication volume is not meaningful discovery/i);
-  assert.match(html, /href="\/signals\/" data-astro-reload(?:="true")?>Signals \//);
-  assert.match(html, /href="\/signals\/science\/" aria-current="location" data-astro-reload/);
+  assert.match(html, /href="\/signals\/">Signals \//);
+  assert.match(html, /href="\/signals\/science\/" aria-current="location"/);
 });
 
 test('science renderer and updater are dependency-free and fail closed', () => {
