+++ ctw.studio/src/components/SiteFooter.astro
@@ -8,8 +8,8 @@
     <nav class="ctw-footer__links" aria-label="Studio destinations">
       <span class="ctw-label">Explore</span>
       <a href="/portfolio/">Projects</a>
-      <a href="/signals/" data-astro-reload>Signals</a>
-      <a href="/nlesc/" data-astro-reload>NLeSC</a>
+      <a href="/signals/">Signals</a>
+      <a href="/nlesc/">NLeSC</a>
     </nav>
     <nav class="ctw-footer__links" aria-label="Contact destinations">
       <span class="ctw-label">Contact</span>
