+++ ctw.studio/src/layouts/GuideLayout.astro
@@ -18,7 +18,6 @@ const { title, description, canonical } = Astro.props;
       description={description}
       canonical={canonical}
       pageCssUrls={['/design-system/guide.css']}
-      transitions={false}
     />
   </head>
   <body class="guide-page">
