+++ ctw.studio/src/layouts/SiteLayout.astro
@@ -35,7 +35,6 @@ const {
       canonical={canonical}
       ogDescription={ogDescription}
       homepageStyles={homepageStyles}
-      transitions
     />
   </head>
   <body class:list={['ctw-scope', bodyClass]}>
@@ -45,7 +44,6 @@ const {
       <slot />
       <SiteFooter />
     </div>
-    <script is:inline data-astro-rerun>window.__ctwFeedbackLoaded = false;</script>
-    <script is:inline src="/feedback.js" data-astro-rerun></script>
+    <script is:inline src="/feedback.js"></script>
   </body>
 </html>
