+++ ctw.studio/src/layouts/WorkshopLayout.astro
@@ -23,7 +23,6 @@ const { title, description, canonical, ogDescription = description, legal = fals
       canonical={canonical}
       ogDescription={ogDescription}
       pageCssUrls={['/tailwind.css']}
-      transitions={false}
     />
     <style is:global>
       :root { color-scheme: dark; }
