+++ ctw.studio/src/pages/index.astro
@@ -16,7 +16,7 @@ import SiteLayout from "../layouts/SiteLayout.astro";
           <div class="ctw-hero__grid">
             <div class="ctw-stack">
               <p class="ctw-section-number">01 / Applied practice</p>
-              <h1 class="ctw-hero__title" id="hero-title" transition:name="portfolio-lead">Applied research <span class="ctw-display--outline">software.</span></h1>
+              <h1 class="ctw-hero__title" id="hero-title">Applied research <span class="ctw-display--outline">software.</span></h1>
               <p class="ctw-lede">Software for research and society. Data infrastructure, ML systems, and tools that create <strong>lasting impact</strong>.</p>
               <div class="ctw-cluster">
                 <a class="ctw-button" href="#contact">Get in touch →</a>
